Deduplicate inline `fn sm_flag` to `src/ipred_prepare.rs`.
