+++ src/recon_tmpl_8.rs
@@ -973,21 +973,7 @@ unsafe extern "C" fn get_uv_inter_txtp(
     return ytxtp;
 }
 use crate::src::msac::dav1d_msac_decode_bools;
-#[inline]
-unsafe extern "C" fn sm_flag(b: *const BlockContext, idx: libc::c_int) -> libc::c_int {
-    if (*b).intra[idx as usize] == 0 {
-        return 0 as libc::c_int;
-    }
-    let m: IntraPredMode = (*b).mode[idx as usize] as IntraPredMode;
-    return if m as libc::c_uint == SMOOTH_PRED as libc::c_int as libc::c_uint
-        || m as libc::c_uint == SMOOTH_H_PRED as libc::c_int as libc::c_uint
-        || m as libc::c_uint == SMOOTH_V_PRED as libc::c_int as libc::c_uint
-    {
-        512 as libc::c_int
-    } else {
-        0 as libc::c_int
-    };
-}
+use crate::src::ipred_prepare::sm_flag;
 #[inline]
 unsafe extern "C" fn sm_uv_flag(
     b: *const BlockContext,
